snap: fix PATH and cmdline options
This commit fixes the LD_LIBRARY_PATH in the pulseaudio
wrapper script, and also updates the pulseaudio command-line
options to account for the fact that the server is running
in system mode.

It also removes the connect* hooks which automatically started
the server when the required interfaces were connected. This
snap is meant to be used on Ubuntu Core and the required snap
interfaces must be granted to the snap via the snap dashboard
page (alsa, and hardware-observe) and by the Snap Store Admins
(audio-playback and audio-record).

NOTE - this last change means that the server needs to be manually
started/enable via snap command or a device-agent using the snapd
REST API.

Signed-off-by: Tony Espy <espy@canonical.com>